+++ src/Resources/contao/dca/tl_news_archive.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  */
 
 $GLOBALS['TL_DCA']['tl_news_archive']['list']['global_operations']['counter_news'] = [
-	'label'      => ($GLOBALS['TL_LANG']['tl_news_archive']['counter_news'] ?? null),
+	'label'      => &$GLOBALS['TL_LANG']['tl_news_archive']['counter_news'],
 	'href'       => 'key=counter',
 	'icon'       => 'bundles/contaocounter/images/counter.png',
 	'attributes' => 'onclick="Backend.getScrollOffset();"',