+++ src/Resources/contao/dca/tl_page.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  */
 
 $GLOBALS['TL_DCA']['tl_page']['list']['global_operations']['counter_pages'] = [
-	'label'      => ($GLOBALS['TL_LANG']['tl_page']['counter_pages'] ?? null),
+	'label'      => &$GLOBALS['TL_LANG']['tl_page']['counter_pages'],
 	'href'       => 'key=counter',
 	'icon'       => 'bundles/contaocounter/images/counter.png',
 	'attributes' => 'onclick="Backend.getScrollOffset();"',